+++ changelog.md
@@ -7,5 +7,9 @@ Added functionality for the cyber contact form.
 # v0.1.6 2026 06 15
 Created the basis for the Am I Being Scammed? questionaire. 
 
-# v0.1.7
+# v0.1.7 2026 06 15
 Finished the logic for the Am I being Scammed? questionaire. 
+
+# v0.1.8 2026 06 15
+- Added the ability to clear selection for each question and a disclaimer at the top of the page.
+- Added a button to clear the entire questionaire as well.